requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
exceptiongroup==1.1.1
fake-useragent==1.1.3
h11==0.14.0
idna==3.4
joblib==1.2.0
mysql-connector-python==8.0.32
numpy==1.24.2
outcome==1.2.0
pandas==1.5.3
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
selenium==3.141.0
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
wsproto==1.2.0
xmltodict==0.13.0
aspose-cells==23.9.0
lxml==4.9.3